@@ -1,21 +1,35 @@
 [project]
 name = "semantiaz"
 version = "0.0.1"
-description = "This is a template repository for Python projects that use uv for their dependency management."
+description = "Comprehensive semantic model management platform for data warehouses with multi-database support, quality assessment, knowledge graph generation, and interactive visualization capabilities."
 authors = [{ name = "Javier Díez Pérez", email = "javier.diez@astrazeneca.com" }]
 readme = "README.md"
-keywords = ['python']
-requires-python = ">=3.9,<4.0"
+keywords = ['semantic-model', 'data-warehouse', 'snowflake', 'knowledge-graph', 'rdf', 'cypher', 'mermaid']
+requires-python = ">=3.10"
 classifiers = [
     "Intended Audience :: Developers",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
+#     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
     "Topic :: Software Development :: Libraries :: Python Modules",
+    "Topic :: Database",
+    "Topic :: Scientific/Engineering :: Information Analysis",
+]
+dependencies = [
+    "click>=8.0.0",
+    "pydantic>=2.0.0",
+    "pyyaml>=6.0",
+    "ibis-framework[postgres,mysql,sqlite,duckdb,bigquery,trino]>=9.0.0",
+    "snowflake-connector-python>=3.0.0",
+    "pandas>=2.0.0",
+    "plotly>=5.0.0",
+    "rdflib>=7.0.0",
+    "fastmcp>=0.1.0",
+    "duckdb>=0.9.0",
 ]
 
 [project.urls]
@@ -56,9 +92,26 @@ show_error_codes = true
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
+python_files = ["test_*.py"]
+python_classes = ["Test*"]
+python_functions = ["test_*"]
+addopts = [
+    "--strict-markers",
+    "--strict-config",
+    "--verbose",
+    "--cov=src/semantiaz",
+    "--cov-report=term-missing",
+    "--cov-report=html",
+    "--cov-fail-under=80"
+]
+markers = [
+    # "slow: marks tests as slow (deselect with '-m "not slow"')",
+    "integration: marks tests as integration tests",
+    "unit: marks tests as unit tests"
+]
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
 line-length = 120
 fix = true
 
